requirements.txt
requests>=2.28.0
PyYAML>=6.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
